Use fsPath instead of path for config file The problem was that a config file path like '/c:/something' was being used instead of the correct path 'c:\something\'. This is the solution suggested by @fluxium.
Unfortunately a number of other changes were
necessary to make the extension build again
with the latest packages on Windows. Then there
were a number of new lint messages checked, and
new messages because of a newer typescript.
I think I got them all fixed except the cypress test
is failing due to some missing monaco element.
This is likely caused by a dom change in a new version
that needs to be accomodated in the test case.